Further reduce size of the state machine box by re-using fields
C# compiler is smart enough to re-use fields iif they have the same type.
Before there was a field for ConfiguredValueTaskAwaitable<byte> and one for ConfiguredValueTaskAwaitable<int>.
Now they got collopsed to just the ConfiguredValueTaskAwaitable<int>-field.

+++ Sally7/RequestExecutor/ConcurrentRequestExecutor.cs
@@ -206,13 +206,13 @@ namespace Sally7.RequestExecutor
         [DebuggerDisplay(nameof(NeedToWait) + ": {" + nameof(NeedToWait) + ",nq}")]
         private class Signal : IDisposable
         {
-            private readonly Channel<byte> channel = Channel.CreateBounded<byte>(1);
+            private readonly Channel<int> channel = Channel.CreateBounded<int>(1);
 
             public void Dispose() => channel.Writer.Complete();
 
             public bool TryInit() => channel.Writer.TryWrite(0);
 
-            public ValueTask<byte> WaitAsync() => channel.Reader.ReadAsync();
+            public ValueTask<int> WaitAsync() => channel.Reader.ReadAsync();
 
             public bool TryRelease() => channel.Writer.TryWrite(0);
